+++ model_utils/__init__.py
@@ -1,4 +1,4 @@
 from .choices import Choices  # noqa:F401
 from .tracker import FieldTracker, ModelTracker  # noqa:F401
 
-__version__ = '3.2.0'
+__version__ = '3.3.0'